Migrated to session.ts.
