+++ server/package.json
@@ -60,6 +60,7 @@
     "winston": "1.0.2"
   },
   "devDependencies": {
+    "@types/node": "^16.11.11",
     "prettier": "^2.2.1"
   }
 }
